+++ adminapi/src/main/java/io/minio/admin/MinioAdminClient.java
@@ -31,7 +31,7 @@ import io.minio.S3Escaper;
 import io.minio.Signer;
 import io.minio.Time;
 import io.minio.admin.messages.DataUsageInfo;
-import io.minio.admin.messages.InfoMessage;
+import io.minio.admin.messages.Message;
 import io.minio.credentials.Credentials;
 import io.minio.credentials.Provider;
 import io.minio.credentials.StaticProvider;
@@ -607,10 +607,10 @@ public class MinioAdminClient {
    * @throws InvalidKeyException thrown to indicate missing of HMAC SHA-256 library.
    * @throws IOException thrown to indicate I/O error on MinIO REST operation.
    */
-  public InfoMessage getServerInfo()
+  public Message getServerInfo()
       throws IOException, NoSuchAlgorithmException, InvalidKeyException {
     try (Response response = execute(Method.GET, Command.INFO, null, null)) {
-      return OBJECT_MAPPER.readValue(response.body().charStream(), InfoMessage.class);
+      return OBJECT_MAPPER.readValue(response.body().charStream(), Message.class);
     }
   }
