+++ adminapi/src/main/java/io/minio/admin/messages/TimedAction.java
@@ -15,30 +15,37 @@
  * limitations under the License.
  */
 
-package io.minio.admin.clusterinfo;
+package io.minio.admin.messages;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.Collections;
-import java.util.Map;
+import java.math.BigDecimal;
 
 /**
- * DiskMetrics has the information about XL Storage APIs
+ * TimedAction contains a number of actions and their accumulated duration in nanoseconds.
  *
- * @see <a href=
- *     "https://github.com/minio/madmin-go/blob/main/info-commands.go#L395">info-commands.go</a>
+ * @see <a href= "https://github.com/minio/madmin-go/blob/main/metrics.go#L244">metrics.go</a>
  */
-public class DiskMetrics {
-  @JsonProperty("lastMinute")
-  private Map<String, TimedAction> lastMinute;
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class TimedAction {
+  @JsonProperty("count")
+  private BigDecimal count;
 
-  @JsonProperty("apiCalls")
-  private Map<String, String> apiCalls;
+  @JsonProperty("acc_time_ns")
+  private BigDecimal accTime;
 
-  public Map<String, TimedAction> lastMinute() {
-    return Collections.unmodifiableMap(lastMinute);
+  @JsonProperty("bytes")
+  private BigDecimal bytes;
+
+  public BigDecimal count() {
+    return count;
   }
 
-  public Map<String, String> apiCalls() {
-    return Collections.unmodifiableMap(apiCalls);
+  public BigDecimal accTime() {
+    return accTime;
+  }
+
+  public BigDecimal bytes() {
+    return bytes;
   }
 }
