+++ adminapi/src/main/java/io/minio/admin/messages/Versions.java
@@ -15,37 +15,30 @@
  * limitations under the License.
  */
 
-package io.minio.admin.clusterinfo;
+package io.minio.admin.messages;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import java.math.BigDecimal;
 
 /**
- * TimedAction contains a number of actions and their accumulated duration in nanoseconds.
+ * Versions contains the number of versions
  *
- * @see <a href= "https://github.com/minio/madmin-go/blob/main/metrics.go#L244">metrics.go</a>
+ * @see <a href=
+ *     "https://github.com/minio/madmin-go/blob/main/info-commands.go#L298">info-commands.go</a>
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class TimedAction {
+public class Versions {
   @JsonProperty("count")
-  private BigDecimal count;
+  private Integer count;
 
-  @JsonProperty("acc_time_ns")
-  private BigDecimal accTime;
+  @JsonProperty("error")
+  private String error;
 
-  @JsonProperty("bytes")
-  private BigDecimal bytes;
-
-  public BigDecimal count() {
+  public Integer count() {
     return count;
   }
 
-  public BigDecimal accTime() {
-    return accTime;
-  }
-
-  public BigDecimal bytes() {
-    return bytes;
+  public String error() {
+    return error;
   }
 }
